CFrame mat4 -> mat4x3 in GLSL
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@1008 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/GLG3D.lib/include/GLG3D/VertexRange.h
@@ -312,7 +312,7 @@ public:
                 size[a] = 0;
             }
 
-            debugAssertM(count[a] == (int)var[a]->m_numElements, 
+            debugAssertM(count[a] == (size_t)var[a]->m_numElements, 
                 "Updated arrays must have the same size they were created with.");
             if (a > 1) {
                 debugAssertM((var[a]->m_pointer == (uint8*)var[a - 1]->m_pointer + size[a - 1]) ||
